+++ R/plot_vpc_exactbins.R
@@ -95,7 +95,7 @@ plot_vpc_exactbins <- function(sim,
   if(!is.null(strat_var)) {check_factor(sim, strat_var)}
   if(!is.null(loq)) {check_numeric_strict(loq)}
 
-  ##Set vpc aesthetics shown ensuring that obsserved points are not plotted by vpc::vpc()
+  ##Set vpc aesthetics shown ensuring that observed points are not plotted by vpc::vpc()
   show_vpc <- list_update(shown,
                        list(obs_dv = TRUE, obs_ci = TRUE,
                             pi = FALSE, pi_as_area = FALSE, pi_ci = TRUE,
@@ -106,9 +106,15 @@ plot_vpc_exactbins <- function(sim,
   #aesthetics for legend based on settings in vpc::new_vpc_theme
   vpctheme <- list_update(theme, pmxhelpr_vpc_theme())
 
-  ##Data Rename
-  sim <- sim |>
-    dplyr::rename((dplyr::any_of(c(output_vars, time_vars))))
+  #Handle Output and Time Variables
+  if(length(unique(c(time_vars[[1]], time_vars[[2]]))) == 2) {
+    sim <- dplyr::rename(sim, dplyr::any_of(c(time_vars, output_vars)))
+  } else {
+    sim <- sim |>
+      dplyr::rename(dplyr::any_of(c(c(NTIME = time_vars[["NTIME"]]),
+                                    output_vars))) |>
+      dplyr::mutate(TIME = NTIME)
+  }
 
   ##Ensure stratification variable is a not an ordered factor (will cause vpc::add_stratification() to fail)
   if(!is.null(strat_var)){