+++ NEWS.md
@@ -1,3 +1,9 @@
+# pmxhelpr 0.3.5
+
+* Add option to pass the same dataset variable to `TIME` and `NTIME` attributes in `time_vars` for all functions
+* Add option to control default aesthetics (linewidth, linetype, size, shape, alpha) in `plot_dvtime` and `plot_popgof`
+* Fix bug in `plot_popgof` that was preventing upper error bars from inheriting the color mapped to "DV"
+
 # pmxhelpr 0.3.4
 
 * Fix bug in `df_mrgsim_replicate` that was precluding use of non-default time variable names